+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix issue where loading a scene containing projectiles or dynamically modified sprites could cause graphical corruption [@untoxa](https://github.com/untoxa)
 - Fix issue where script event title would show wrong local variable name for scenes [@pau-tomas](https://github.com/pau-tomas)
 - Fix muting bug with FXH parser [@coffeevalenbat](https://github.com/coffeevalenbat)
+- Fix issue where Animation State value for projectiles was being ignored
 
 ### Performance